Refactored code and updated readme
