+++ .travis.yml
@@ -13,10 +13,7 @@
 # limitations under the License.
 
 script:
-  - wget https://product-dist.ballerina.io/downloads/1.0.0/ballerina-linux-installer-x64-1.0.0.deb
-  - sudo dpkg -i ballerina-linux-installer-x64-1.0.0.deb
+  - wget https://product-dist.ballerina.io/downloads/1.0.1/ballerina-linux-installer-x64-1.0.1.deb
+  - sudo dpkg -i ballerina-linux-installer-x64-1.0.1.deb
   - sudo apt-get install -f
   - mvn clean install -DskipTests=true
-  - cd ftp
-  - ballerina --version
-  - ballerina build -c --skip-tests ftp
