--- compiler-plugin/src/main/resources/META-INF/services/org.ballerinalang.compiler.plugins.CompilerPlugin
@@ -1,1 +1,0 @@
-org.wso2.ei.module.ftp.ServiceCompilerPlugin
