--- ftp-utils/src/main/assembly/dist.xml
@@ -1,28 +1,0 @@
-<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.0 http://maven.apache.org/xsd/assembly-1.1.0.xsd">
-    <id>bin</id>
-    <includeBaseDirectory>false</includeBaseDirectory>
-    <formats>
-        <format>zip</format>
-    </formats>
-    <files>
-        <file>
-            <source>${project.build.directory}/${project.artifactId}-module-${project.version}.jar</source>
-            <outputDirectory>./dependencies</outputDirectory>
-            <fileMode>0644</fileMode>
-        </file>
-    </files>
-    <fileSets>
-        <fileSet>
-            <directory>${project.build.directory}/generated-balo/repo</directory>
-            <outputDirectory>./balo</outputDirectory>
-            <fileMode>0644</fileMode>
-        </fileSet>
-        <fileSet>
-            <directory>${project.build.directory}/classes/installation-scripts</directory>
-            <outputDirectory>.</outputDirectory>
-            <fileMode>0755</fileMode>
-        </fileSet>
-    </fileSets>
-</assembly>
