--- ftp-utils/src/main/resources/installation-scripts/README.txt
@@ -1,10 +1,0 @@
-Instructions for usage
-
-This guide contains instructions to install the FTP Ballerina module into your Ballerina installation. This is a one-time task.
-Once done, you can then use this module across all your Ballerina projects.
-
-In this directory, you can find two scripts of interest for you. Please note that when running the following scripts,
-you will need write permission to Ballerina installation location.
-
-1. install.{sh/bat} - Installs the FTP module.
-2. uninstall.{sh/bat} - Uninstalls the FTP module if installed already.
